WIP: Trying to optimise info messages in QT
At the moment, calls to append text to the output pane involve messages
that use threads. This is possibly ineffecient and is a problem on RPis
for code which produces lots of output. This branch is a work in
progress trying to minimize calls to append and setting color properties
by building up a string for use with insertHtml. At the moment this
doesn't print line breaks or use the right colors.