WIP: Add LADSPA plugins to Sonic Pi
This is just on a branch to keep track of my work so far.

The good news - with a little work, you can use LADSPA plugins from
Sonic Pi.

The bad news - getting this ready for distribution is going to be hard.
It would involve changes upstream in Overtone (adding LADSPA UGen which,
unlike any other UGen I can think of, has a variable number of args) and
also bundling the compiled LADSPA binaries for the three platforms. We
could potentially steal the compiled versions from LMMS
(https://lmms.io) but I'd want to check they were cool with that.

This would add a huge number of effects to Sonic Pi, some of them that
aren't available in any other form that I'm aware of. The Tube Warmth
that I've implemented here is a good example of that.