WIP: Now compiling on OSX
These changes allowed me to compile the scope branch on OS X. The
main stumbling block was the use of clang which defaults to use a newer
version of the C++ library called `libc++` - this conflicts with the
older, more widely used library `libstdc++`.

These two don't play nicely together so every dependency of the Qt GUI
(boost, qwt, qscintilla and Qt itself)
has to be compiled with either `libc++` or `libstdc++` but they can't
mix and match. This results in errors like:

```
Undefined symbols for architecture x86_64:
  "std::__1::__vector_base_common<true>::__throw_length_error() const",
referenced from:
      std::__1::vector<SonicPiLog::Message,
std::__1::allocator<SonicPiLog::Message>
>::vector(std::__1::vector<SonicPiLog::Message,
>std::__1::allocator<SonicPiLog::Message> > const&) in main.o
```

or

```
Undefined symbols for architecture x86_64:
  "std::__1::__vector_base_common<true>::__throw_length_error() const",
referenced from:
```

The way to check which lib is being used is to run `otool -L` on the
compiled app or lib. `otool -L` is the mac equivalent of `ldd`. Running
that produces something like this

```
$ otool -L lib/qwt.framework/Versions/6/qwt
lib/qwt.framework/Versions/6/qwt:
        qwt.framework/Versions/6/qwt (compatibility version 6.1.0, current version 6.1.2)
        /usr/local/opt/qt5/lib/QtPrintSupport.framework/Versions/5/QtPrintSupport (compatibility version 5.4.0, current version 5.4.1)
        /usr/local/opt/qt5/lib/QtWidgets.framework/Versions/5/QtWidgets (compatibility version 5.4.0, current version 5.4.1)
        /usr/local/opt/qt5/lib/QtGui.framework/Versions/5/QtGui (compatibility version 5.4.0, current version 5.4.1)
        /usr/local/opt/qt5/lib/QtCore.framework/Versions/5/QtCore (compatibility version 5.4.0, current version 5.4.1)
        /System/Library/Frameworks/DiskArbitration.framework/Versions/A/DiskArbitration (compatibility version 1.0.0, current version 1.0.0)
        /System/Library/Frameworks/IOKit.framework/Versions/A/IOKit (compatibility version 1.0.0, current version 275.0.0)
        /usr/local/opt/qt5/lib/QtSvg.framework/Versions/5/QtSvg (compatibility version 5.4.0, current version 5.4.1)
        /usr/local/opt/qt5/lib/QtOpenGL.framework/Versions/5/QtOpenGL (compatibility version 5.4.0, current version 5.4.1)
        /usr/local/opt/qt5/lib/QtConcurrent.framework/Versions/5/QtConcurrent (compatibility version 5.4.0, current version 5.4.1)
        /System/Library/Frameworks/OpenGL.framework/Versions/A/OpenGL (compatibility version 1.0.0, current version 1.0.0)
        /System/Library/Frameworks/AGL.framework/Versions/A/AGL (compatibility version 1.0.0, current version 1.0.0)
        /usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 120.0.0)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1213.0.0)
```

Note the line `/usr/lib/libc++.1.dylib`. If you see `libstdc++.6.dylib`
you are using a lib that probably hasn't been compiled with clang. I
opted to make all the deps `libc++` but they could easily work with the
other, it just needs to be consistent.